fix(examples): vendor shared serdes so child-context examples bundle correctly
The large-payload and virtual serdes examples imported uppercaseSerdes from
run-in-child-context/serdes/run-in-child-context-serdes, which is itself a
catalog example (its own rollup entry chunk). rollup emitted an unresolved
require('./run-in-child-context-serdes.js') into the dependent bundles; since
package.js only ships the handler bundle plus vendors.js, the deployed function
failed at runtime with "Cannot find module './run-in-child-context-serdes.js'".

Move uppercaseSerdes into examples/shared/uppercase-serdes.ts. The example
catalog scanner explicitly skips a top-level shared/ directory, and the module
exports no handler/config, so it is not a catalog entry. rollup folds it into
the shared vendors chunk that is packaged with every handler.

Verified against the built artifacts: the large-payload and virtual bundles now
require only ./vendors.js (no sibling-handler import), and require()-ing a
bundled handler with just vendors.js alongside loads successfully.
